Image collection tool changes

+++ tools/ImageCollectionTool/src/main/java/org/openimaj/tools/imagecollection/processor/DirectoryImageProcessor.java
@@ -22,18 +22,13 @@ public class DirectoryImageProcessor<T extends Image<?, T>> extends ImageCollect
 
 	@Override
 	public void start() throws IOException{
-		if(this.directoryFile.isDirectory() ){
-			if(this.directoryFile.exists()){
-				if(force)
-					FileUtils.deleteRecursive(this.directoryFile);
-				else
-					throw new IOException("Directory already exists");
+		if(this.directoryFile.exists()){
+			if(this.directoryFile.isDirectory() ){
+				if(force) FileUtils.deleteRecursive(this.directoryFile);
 			}
-			
-			
-			if(!this.directoryFile.mkdirs()){
-				throw new IOException("Can't create directory");
-			}
+		}
+		if(!this.directoryFile.mkdirs()){
+			throw new IOException("Can't create directory");
 		}
 		// Directory should exist and be a directory now
 	}
